refactor: move overlay permission functions to core package
