+++ app-service-hilt/src/main/java/com/github/only52607/compose/window/hilt/MainActivity.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.github.only52607.compose.window.checkOverlayPermission
+import com.github.only52607.compose.core.checkOverlayPermission
 import com.github.only52607.compose.window.hilt.repository.UserPreferencesRepository
 import com.github.only52607.compose.window.hilt.ui.DialogPermission
 import com.github.only52607.compose.window.hilt.ui.theme.ComposeFloatingWindowTheme
