+++ app-service/src/main/java/com/github/only52607/compose/window/service/ui/DialogPermission.kt
@@ -13,8 +13,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.compose.LocalLifecycleOwner
-import com.github.only52607.compose.window.checkOverlayPermission
-import com.github.only52607.compose.window.requestOverlayPermission
+import com.github.only52607.compose.core.checkOverlayPermission
+import com.github.only52607.compose.core.requestOverlayPermission
 import com.github.only52607.compose.window.service.R
 
 @Composable
