+++ library/src/main/java/com/github/only52607/compose/core/CoreFloatingWindow.kt
@@ -19,7 +19,6 @@ import androidx.savedstate.SavedStateRegistry
 import androidx.savedstate.SavedStateRegistryController
 import androidx.savedstate.SavedStateRegistryOwner
 import com.github.only52607.compose.service.SERVICE_TAG
-import com.github.only52607.compose.window.requestOverlayPermission
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.SupervisorJob
