+++ library/src/main/java/com/github/only52607/compose/core/Permissions.kt
@@ -1,4 +1,4 @@
-package com.github.only52607.compose.window
+package com.github.only52607.compose.core
 
 import android.content.Context
 import android.content.Intent
